Refactor using Renderers
