+++ src/main/scala/com/fortysevendeg/scala/android/ui/materiallist/MaterialListActivity.scala
@@ -1,30 +1,17 @@
 package com.fortysevendeg.scala.android.ui.materiallist
 
 import android.os.Bundle
-import android.support.design.widget.Snackbar
 import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget._
 import android.view.MenuItem
-import com.fortysevendeg.macroid.extras.RecyclerViewTweaks._
-import com.fortysevendeg.scala.android.{R, TR, TypedFindView}
-import macroid.{Ui, Contexts}
+import com.fortysevendeg.scala.android.{R, TypedFindView}
+import macroid.Contexts
 import macroid.FullDsl._
 
 class MaterialListActivity
   extends AppCompatActivity
   with TypedFindView
-  with Transformations
+  with ScreenRenderer
   with Contexts[AppCompatActivity] {
-
-  lazy val content = Option(findView(TR.content))
-
-  lazy val toolBar = Option(findView(TR.toolbar))
-
-  lazy val appBarLayout = Option(findView(TR.app_bar_layout))
-
-  lazy val recycler = Option(findView(TR.recycler))
-
-  lazy val fabActionButton = Option(findView(TR.fab_action_button))
 
   override def onCreate(savedInstanceState: Bundle) = {
     super.onCreate(savedInstanceState)
@@ -34,24 +21,7 @@ class MaterialListActivity
     toolBar foreach  setSupportActionBar
     getSupportActionBar.setDisplayHomeAsUpEnabled(true)
 
-    val layoutManager = new GridLayoutManager(this, 2)
-
-    runUi(
-      (fabActionButton
-          <~ fabIcon
-          <~ On.click {
-            Ui {
-              content foreach {
-                Snackbar.make(_, getString(R.string.material_list_add_item), Snackbar.LENGTH_LONG).show()
-              }
-            }
-          }) ~
-          (recycler
-              <~ rvAdapter(new ImageListAdapter())
-              <~ rvFixedSize
-              <~ rvLayoutManager(layoutManager))
-    )
-
+    runUi(init)
   }
 
   override def onOptionsItemSelected(item: MenuItem): Boolean = item.getItemId match {
